feat(internal/librariangen): initial implementation of configure
This is completely lacking in tests, which should be provided before
merging.

The implementation will currently require a playbook along the lines of:

1. Run `librarian generate -api=google/cloud/abc/v1 -library=abc`
1. Run `(cd abc && go mod tidy)`
1. Run `go work use ./abc`
1. Commit changes and push

Steps 2 and 3 are only needed for whole new libraries (as opposed to
new APIs within existing libraries)

It would be nice to be able to get rid of steps 2 and 3 entirely, but
they're difficult.

We can't run `go mod tidy` without source code,
which either means running the generator during configure (which is
ugly and somewhat against the spirit of configure) or do so in
`generate`, but *only* on the very first run.

We can't run `go work use` without copying *all* of the `go.mod` files
referenced in there into the output directory (and then removing
them again). We could modify `go.work` directly, but that's ugly too.

These can be addressed in the longer term as we attempt to get to
fully automated onboarding, but having the steps in the playbook
introduces *relatively* little toil, while saving a *lot* of work in the
container.