fix(storage): Fix takeover response handling.
A takeover response can return a handle. In that case, a retry should
present that handle. Otherwise it will attempt a takeover when it
retries. After a new takeover, the client can't tell if a partial write
after the last known flush offset was performed by itself, or by some
other writer. Therefore the client can't tell if it's safe to retry from
the last known flush offset, and the server will return an error if it
repeats offsets.

The write handle allows the client to present enough information to the
server to correctly identify replays on the same write session, avoiding
an error when the client repeats bytes that are already durable.