feat: Support scopes field from impersonated credential json (#13308)
If impersonated json contains scopes field and the user does not provide
scopes through options, then use the scopes field value for the
impersonated token.

Hard code the scopes for source credential to be always
`cloud-platform`. Unless we do this, if user provides a scope other than
cloud-platform or IAM, then the call to iamcredentials will fail. (This
is an existing bug in the code)

Sample JSON:
```
{
  "delegates": [],
  "scopes": [
    "https://www.googleapis.com/auth/drive"
  ],
  "service_account_impersonation_url": "https://iamcredentials.googleapis.com/v1/projects/-/serviceAccounts/service-account-email@project-name.iam.gserviceaccount.com:generateAccessToken",
  "source_credentials": {
    "client_id": "oauth_client_id",
    "client_secret": "oauth_client_secret",
    "refresh_token": "user_refresh_token",
    "type": "authorized_user",
    "universe_domain": "googleapis.com"
  },
  "type": "impersonated_service_account"
}
```