fix(librariangen): don't run tests from librariangen (#13336)
This will fix automated generation problems for now. There are at least
two issues:

- https://github.com/googleapis/librarian/issues/2821 is about go test
failing when downloading dependencies. It's possible that running "go
mod download" before "go test" fixes that, but it would be good to know
more.
- https://github.com/googleapis/google-cloud-go/issues/13335 is about
our tests run with `-short` not being as unit-testy as we might want.

Note that we still build the production code, which should protect us
against generators creating odd code - and the tests will still run on
GitHub PRs as normal.