chore(internal/librariangen): remove the legacy protoc/gRPC plugins (#13372)
Fixes https://github.com/googleapis/librarian/issues/2796

Note that we *may* need to reinstate this in the Dockerfile in order to
support go-genproto, but we shouldn't need to reinstate the logic in the
Bazel parser etc.

We deliberately spot cases where we *would* have used the old plugin, in
order to fail hard instead of just omitting the gRPC output.

This change affects which version of protoc is used for proto-only
paths, but running "librarian update-image" using a locally-build
container demonstrates that this doesn't affect the generated code.