testing(bigquery): remove remote UDF test (#13473)
This test relied on a fake cloud run app ID, and special logic in the
backend allowed this to proceed even with the invalid app.  Now that
testing has moved to a new region validity checks prevent this test
from proceeding because the app does not exist.

This PR removes the test and accompanying utility functionality.