perf(spanner): only create sessions if multiplexed sessions are disabled (#13477)
Only create regular sessions if multiplexed sessions have not been
enabled for all operations. If multiplexed sessions have been enabled
for all operations, and the application has not set a custom number of
minimum number of sessions, then the session pool will no longer create
a default of 100 regular sessions. These then also do not need to be
maintained by the pool.

This change will cause the session pool to contain 0 session by default.
The session pool will only be initialized with a non-zero number of
sessions if any of the following is true:
1. `MinOpened` has been set to a non-default value (meaning: `MinOpened
!= 100`)
2. Multiplexed sessions have been disabled for at least one type of
operation.