test(bigquery/v2/query): deflake TestIntegration_QueryCancelWait
This PR adjusts test expectations to be more precise with the concurrent
execution semantics.  Rather than leaning on a generate_array query
to slow execution, this PR switches to using a script that has a known
min duration, and switches to use a WaitGroup to test the premature
cancellation behavior.